+++ CHANGES.md
@@ -1,3 +1,10 @@
+## Version 0.1.2, 2017.10.30
+
+* Add default footer for docs [#7](https://github.com/NotNinja/brander/issues/7) [#12](https://github.com/NotNinja/brander/issues/12)
+* Add scale option for SVG conversions [#10](https://github.com/NotNinja/brander/issues/10) [#15](https://github.com/NotNinja/brander/issues/15)
+* Add baseFile and baseUrl options for SVG conversions [#11](https://github.com/NotNinja/brander/issues/11) [#15](https://github.com/NotNinja/brander/issues/15)
+* Add quiet option to CLI [#14](https://github.com/NotNinja/brander/issues/14) [#17](https://github.com/NotNinja/brander/issues/17)
+
 ## Version 0.1.1, 2017.10.24
 
 * Title not rendered for each file group within asset-feature document [#2](https://github.com/NotNinja/brander/issues/2)